[Chore] 파일 분리
