+++ Unsplash/View/MainPhotoView/View/MainPhotoViewController.swift
@@ -5,7 +5,6 @@
 //  Created by 김수경 on 2024/03/07.
 //
 
-import Foundation
 import UIKit
 
 final class MainPhotoViewController: UIViewController {
