+++ Unsplash/View/MainPhotoView/View/RecentCollectionViewLayout.swift
@@ -1,0 +1,96 @@
+//
+//  RecentCollectionViewLayout.swift
+//  Unsplash
+//
+//  Created by 김수경 on 2024/03/15.
+//
+
+import UIKit
+
+final class RecentCollectionViewLayout: UICollectionViewLayout {
+    
+    weak var delegate: RecentCollectionViewLayoutDelegate?
+    
+    private let numberOfColumns = 2
+    private let cellPadding: CGFloat = 4
+
+    private var attributeCache: [UICollectionViewLayoutAttributes] = []
+    private var contentHeight: CGFloat = 0
+    
+    private var contentWidth: CGFloat {
+        guard let collectionView = collectionView else {
+            return 0
+        }
+        
+        let insets = collectionView.contentInset
+        
+        return collectionView.bounds.width - (insets.left + insets.right)
+    }
+}
+
+
+// MARK: - override method
+
+extension RecentCollectionViewLayout {
+    
+    override var collectionViewContentSize: CGSize {
+        return CGSize(width: contentWidth, height: contentHeight)
+    }
+    
+    override func prepare() {
+        guard attributeCache.isEmpty == true else { return }
+      
+        guard let collectionView = collectionView else { return }
+        
+        let columnWidth = contentWidth / CGFloat(numberOfColumns)
+        
+        var xOffset: [CGFloat] = []
+        
+        for column in 0..<numberOfColumns {
+            xOffset.append(CGFloat(column) * columnWidth)
+        }
+        
+        var column = 0
+        var yOffset: [CGFloat] = .init(repeating: 0, count: numberOfColumns)
+        
+        for item in 0..<collectionView.numberOfItems(inSection: 0) {
+            let indexPath = IndexPath(item: item, section: 0)
+            
+            let photoHeight = delegate?.collectionView(collectionView,
+                                                       heightForPhotoAtIndexPaht: indexPath) ?? columnWidth
+            let height = cellPadding * 2 + photoHeight
+            let frame = CGRect(x: xOffset[column],
+                               y: yOffset[column],
+                               width: columnWidth,
+                               height: height)
+            let insetFrame = frame.insetBy(dx: cellPadding, dy: cellPadding)
+            
+            let attributes = UICollectionViewLayoutAttributes(forCellWith: indexPath)
+            attributes.frame = insetFrame
+            
+            attributeCache.append(attributes)
+            
+            contentHeight = max(contentHeight, frame.maxY)
+            yOffset[column] = yOffset[column] + height
+            
+            column = column < (numberOfColumns - 1) ? (column + 1) : 0
+        }
+    }
+    
+    override func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]? {
+        var visibleLayoutAttributes: [UICollectionViewLayoutAttributes] = []
+
+        for attributes in attributeCache {
+            if attributes.frame.intersects(rect) {
+                visibleLayoutAttributes.append(attributes)
+            }
+        }
+        
+        return visibleLayoutAttributes
+    }
+    
+    override func layoutAttributesForItem(at indexPath: IndexPath) -> UICollectionViewLayoutAttributes? {
+        attributeCache[indexPath.item]
+    }
+    
+}
